+++ sdks/recached-react/src/index.ts
@@ -1,4 +1,4 @@
 export { RecachedProvider, useRecached } from './context';
-export { useKey, useKeyJSON } from './useKey';
+export { useKey, useKeyBytes, useKeyJSON } from './useKey';
 export { useKeys, type KeyValuePair } from './useKeys';
 export { usePubSub } from './usePubSub';